+++ local-tests/tests/wrapped-keys/testImportWrappedKey.ts
@@ -3,10 +3,8 @@ import { TinnyEnvironment } from 'local-tests/setup/tinny-environment';
 import { api } from '@lit-protocol/wrapped-keys';
 import { getPkpSessionSigs } from 'local-tests/setup/session-sigs/get-pkp-session-sigs';
 import { randomSolanaPrivateKey } from 'local-tests/setup/tinny-utils';
-import { listPrivateKeyMetadata } from '../../../packages/wrapped-keys/src/lib/service-client';
-import { getFirstSessionSig } from '../../../packages/wrapped-keys/src/lib/utils';
 
-const { importPrivateKey } = api;
+const { importPrivateKey, listEncryptedKeyMetadata } = api;
 
 /**
  * Test Commands:
@@ -44,9 +42,9 @@ export const testImportWrappedKey = async (devEnv: TinnyEnvironment) => {
       );
     }
 
-    const keys = await listPrivateKeyMetadata({
-      sessionSig: getFirstSessionSig(pkpSessionSigs),
-      litNetwork: devEnv.litNodeClient.config.litNetwork,
+    const keys = await listEncryptedKeyMetadata({
+      pkpSessionSigs,
+      litNodeClient: devEnv.litNodeClient,
     });
 
     if (keys.length !== 1 || keys[0].id !== id) {
