+++ packages/wrapped-keys/src/lib/service-client/types.ts
@@ -8,10 +8,11 @@ interface BaseApiParams {
 }
 
 export type FetchKeyParams = BaseApiParams & {
+  pkpAddress: string;
   id: string;
 };
 
-export type ListKeysParams = BaseApiParams;
+export type ListKeysParams = BaseApiParams & { pkpAddress: string };
 
 export type SupportedNetworks = Extract<
   LIT_NETWORKS_KEYS,
@@ -21,17 +22,11 @@ export type SupportedNetworks = Extract<
 export interface StoreKeyParams extends BaseApiParams {
   storedKeyMetadata: Pick<
     StoredKeyData,
-    | 'pkpAddress'
-    | 'publicKey'
-    | 'keyType'
-    | 'dataToEncryptHash'
-    | 'ciphertext'
-    | 'memo'
+    'publicKey' | 'keyType' | 'dataToEncryptHash' | 'ciphertext' | 'memo'
   >;
 }
 
 export interface StoreKeyBatchParams extends BaseApiParams {
-  pkpAddress: string;
   storedKeyMetadataBatch: Pick<
     StoredKeyData,
     'publicKey' | 'keyType' | 'dataToEncryptHash' | 'ciphertext' | 'memo'
